Piece listing and glass cup tube dispensers
We got a list of piece info which gets turned into Ui buttons which you can then click on and place down the piece and it takes away from your budget.
Made oranges smaller and made glass cups.
Made a debug dispenser, I`ll make a tube model and spout next time.
Gotta work on making the base gameplay loop work though.
So next things to work on, is something to let you juice oranges and another thing to let you hold the cup in place for the juice to fall into.
Then for the cup to go onto the target spot and have it finish the level.
Level result screen too.